Move file opener to ui-components
