+++ packages/desktop-gui/src/main.scss
@@ -4,4 +4,4 @@
 @import 'styles/components/*';
 @import '!(styles)*/**/*';
 @import '../../ui-components/src/dropdown';
-@import '../../ui-components/src/editor-picker';
+@import '../../ui-components/src/opener/opener';
