+++ packages/reporter/src/main-runner.scss
@@ -5,5 +5,5 @@
 @import 'lib/tooltip';
 @import 'lib/modal';
 @import '../../../node_modules/@reach/dialog/styles.css';
-@import '../../ui-components/src/editor-picker';
+@import '../../ui-components/src/opener/opener';
 @import './!(lib)*/**/*';
