--- packages/reporter/src/opener/file-model.ts
@@ -1,7 +1,0 @@
-export interface FileDetails {
-    absoluteFile: string
-    column: number
-    line: number
-    originalFile: string
-    relativeFile: string
-}
