+++ packages/reporter/src/opener/file-opener.tsx
@@ -1,16 +1,10 @@
-import _ from 'lodash'
-import { action } from 'mobx'
-import { observer, useLocalStore } from 'mobx-react'
-import React, { MouseEvent } from 'react'
+import { observer } from 'mobx-react'
+import React from 'react'
 
-import EditorPickerModal, { Editor } from './editor-picker-modal'
-import { FileDetails } from './file-model'
 import events from '../lib/events'
 
-interface GetUserEditorResult {
-  preferredOpener?: Editor
-  availableEditors?: Editor[]
-}
+//@ts-ignore
+import { GetUserEditorResult, Editor, FileDetails, FileOpener as Opener } from '@packages/ui-components'
 
 interface Props {
   fileDetails: FileDetails,
@@ -26,68 +20,22 @@ const openFile = (where: Editor, { absoluteFile: file, line, column }: FileDetai
   })
 }
 
-const FileOpener = observer(({ fileDetails, className }: Props) => {
-  const state = useLocalStore(() => ({
-    editors: [] as Editor[],
-    chosenEditor: {} as Editor,
-    isLoadingEditor: false,
-    isModalOpen: false,
-    setChosenEditor: action((editor: Editor) => {
-      state.chosenEditor = editor
-    }),
-    setEditors: action((editors: Editor[]) => {
-      state.editors = editors
-    }),
-    setIsLoadingEditor: action((isLoading: boolean) => {
-      state.isLoadingEditor = isLoading
-    }),
-    setIsModalOpen: action((isOpen: boolean) => {
-      state.isModalOpen = isOpen
-    }),
-  }))
+const getUserEditor = (callback: (result: GetUserEditorResult) => any) => {
+  events.emit('get:user:editor', callback)
+}
 
-  const attemptOpenFile = (e: MouseEvent) => {
-    e.preventDefault()
+const setUserEditor = (editor: Editor) => {
+  events.emit('set:user:editor', editor)
+}
 
-    if (state.isLoadingEditor) return
-
-    state.setIsLoadingEditor(true)
-
-    // TODO: instead of the back-n-forth, send 'open:file' or similar, and if the
-    // user editor isn't set, it should send back the available editors
-    events.emit('get:user:editor', (result: GetUserEditorResult) => {
-      state.setIsLoadingEditor(false)
-
-      if (result.preferredOpener) {
-        return openFile(result.preferredOpener, fileDetails)
-      }
-
-      state.setEditors(result.availableEditors || [])
-      state.setIsModalOpen(true)
-    })
-  }
-
-  const setEditor = (editor: Editor) => {
-    events.emit('set:user:editor', editor)
-    state.setIsModalOpen(false)
-    state.setChosenEditor({} as Editor)
-    openFile(editor, fileDetails)
-  }
-
-  const { originalFile, line, column } = fileDetails
-
+const FileOpener = observer((props: Props) => {
   return (
-    <a className={className} onClick={attemptOpenFile} href='#'>
-      {originalFile}{!!line && `:${line}`}{!!column && `:${column}`}
-      <EditorPickerModal
-        chosenEditor={state.chosenEditor}
-        editors={state.editors}
-        isOpen={state.isModalOpen}
-        onSetEditor={setEditor}
-        onSetChosenEditor={state.setChosenEditor}
-        onClose={_.partial(state.setIsModalOpen, false)}
-      />
-    </a>
+    <Opener
+      openFile={openFile}
+      getUserEditor={getUserEditor}
+      setUserEditor={setUserEditor}
+      {...props}
+    />
   )
 })
