+++ packages/ui-components/cypress/support/test-entry.scss
@@ -4,4 +4,4 @@
 @import "../../node_modules/@fortawesome/fontawesome-free/scss/fontawesome.scss";
 
 @import '../../src/dropdown';
-@import '../../src/editor-picker';
+@import '../../src/opener/opener';
