+++ packages/ui-components/src/.eslintrc.json
@@ -2,5 +2,25 @@
   "extends": [
     "plugin:@cypress/dev/react",
     "plugin:@cypress/dev/tests"
-  ]
+  ],
+  "parser": "@typescript-eslint/parser",
+  "settings": {
+    "react": {
+      "version": "16.12"
+    }
+  },
+  "rules": {
+    "arrow-body-style": "off",
+    "no-unused-vars": "off",
+    "react/jsx-filename-extension": [
+      "warn",
+      {
+        "extensions": [
+          ".js",
+          ".jsx",
+          ".tsx"
+        ]
+      }
+    ]
+  }
 }
