+++ packages/ui-components/src/index.jsx
@@ -2,6 +2,12 @@ export { default as BrowserIcon } from './browser-icon'
 
 export { default as Dropdown } from './dropdown'
 
-export { default as EditorPicker } from './editor-picker'
+export { default as EditorPicker } from './opener/editor-picker'
+
+export { default as EditorPickerModal } from './opener/editor-picker-modal'
+
+export { default as FileOpener } from './opener/file-opener'
+
+export * from './opener/file-model'
 
 export * from './select'
