+++ packages/ui-components/src/opener/editor-picker-modal.tsx
@@ -9,14 +9,8 @@ import cs from 'classnames'
 import React from 'react'
 import VisuallyHidden from '@reach/visually-hidden'
 // @ts-ignore
-import { EditorPicker } from '@packages/ui-components'
-
-export interface Editor {
-  id: string
-  name: string
-  openerId: string
-  isOther: boolean
-}
+import EditorPicker from './editor-picker'
+import { Editor } from './file-model'
 
 interface Props {
   chosenEditor: Editor
