+++ packages/ui-components/src/opener/opener.scss
@@ -1,4 +1,4 @@
-@import './select/select';
+@import '../select/select';
 
 .editor-picker {
   label {
@@ -32,5 +32,29 @@
     display: block;
     padding-left: 5.2em;
     width: 100%;
+  }
+}
+
+.editor-picker-modal {
+  max-width: 40em;
+
+  .editor-picker {
+    margin-bottom: 1em;
+  }
+
+  .controls {
+    > span:first-child {
+      order: 1;
+    }
+
+    button.is-disabled,
+    button.is-disabled:hover,
+    button.is-disabled:focus {
+      background: $pass !important;
+      cursor: default !important;
+      opacity: 0.5;
+    }
+
+    padding: 1em 1em 1em;
   }
 }
